requirements.txt
accelerate==0.31.0
anthropic==0.28.0
gpt_batch==0.1.6
numpy==1.26.4
openai==1.34.0
protobuf==5.27.1
python-dotenv==1.0.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
vertexai==1.49.0
vllm==0.5.0